+++ src/unpoly-migrate/syntax.js
@@ -28,3 +28,88 @@ up.migrate.targetMacro = function(queryAttr, fixedResultAttrs, callback) {
     callback?.()
   })
 }
+
+/*-
+Registers a function to be called when an element with
+the given selector is inserted into the DOM. The function is called
+with each matching element as a
+[jQuery object](https://learn.jquery.com/using-jquery-core/jquery-object/).
+
+If you're not using jQuery, use `up.compiler()` instead, which calls
+the compiler function with a native element.
+
+### Example
+
+This jQuery compiler will insert the current time into a
+`<div class='current-time'></div>`:
+
+```js
+up.$compiler('.current-time', function($element) {
+  var now = new Date()
+  $element.text(now.toString())
+})
+```
+
+@function up.$compiler
+@param {string} selector
+  The selector to match.
+@param {Object} [options]
+  See [`options` argument for `up.compiler()`](/up.compiler#parameters).
+@param {Function($element, data)} compiler
+  The function to call when a matching element is inserted.
+
+  See [`compiler` argument for `up.compiler()`](/up.compiler#parameters).
+@deprecated
+  Use `up.compiler()` instead.
+*/
+up.$compiler = function(selector, ...definitionArgs) {
+  let $fn = definitionArgs.pop()
+
+  up.compiler(...definitionArgs, function(element, ...fnArgs) {
+    let $element = jQuery(element)
+    return $fn($element, ...fnArgs)
+  })
+}
+
+
+/*-
+Registers a [compiler](/up.compiler) that is run before all other compilers.
+The compiler function is called with each matching element as a
+[jQuery object](https://learn.jquery.com/using-jquery-core/jquery-object/).
+
+If you're not using jQuery, use `up.macro()` instead, which calls
+the macro function with a native element.
+
+### Example
+
+```js
+up.$macro('[content-link]', function($link) {
+  $link.attr(
+    'up-target': '.content',
+    'up-transition': 'cross-fade',
+    'up-duration':'300'
+  )
+})
+```
+
+@function up.$macro
+@param {string} selector
+  The selector to match.
+@param {Object} options
+  See [`options` argument for `up.compiler()`](/up.compiler#parameters).
+@param {Function(element, data)} macro
+  The function to call when a matching element is inserted.
+
+  See [`compiler` argument for `up.compiler()`](/up.compiler#parameters).
+@deprecated
+  Use `up.macro()` instead.
+*/
+up.$macro = function(selector, ...definitionArgs) {
+  let $fn = definitionArgs.pop()
+
+  up.macro(...definitionArgs, function(element, ...fnArgs) {
+    let $element = jQuery(element)
+    return $fn($element, ...fnArgs)
+  })
+}
+